Global update - php
Part 1 of 2

I have not placed a commit in a long time, I have made many changes over the past year. I have added features across the entire project, too many to comment on. This portion of the commit contains the php version of this site which has been scrapped for a flask/python version of the same project.